# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Azure SQL TCO API
  version: 0.1.0
  description: >-
    Server-authoritative API for estimating AWS and on-premises SQL Server costs
    against Azure SQL Managed Instance. Results are estimates, not quotes.
servers:
  - url: /api/v1
...
paths:
...
  /catalog/aws/ec2/instances:
    get:
      operationId: listEc2Instances
      tags: [Catalogs]
      parameters:
        - $ref: '#/components/parameters/AwsRegion'
      responses:
        '200':
          description: EC2 instance metadata available in the latest regional snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ec2InstanceCatalog'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/aws/rds/instances:
    get:
      operationId: listRdsInstances
      tags: [Catalogs]
      parameters:
        - $ref: '#/components/parameters/AwsRegion'
      responses:
        '200':
          description: RDS instance metadata available in the latest regional snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RdsInstanceCatalog'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/aws/rds/options:
    get:
      operationId: listRdsOptions
      tags: [Catalogs]
      parameters:
        - $ref: '#/components/parameters/AwsRegion'
        - name: instance_type
          in: query
          required: true
          schema:
            type: string
        - name: deployment
          in: query
          required: true
          schema:
            type: string
            enum: [single_az, multi_az]
      responses:
        '200':
          description: Commercial and storage options available for the selected RDS shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RdsOptionCatalog'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/aws/ebs/types:
    get:
      operationId: listEbsTypes
      tags: [Catalogs]
      parameters:
        - $ref: '#/components/parameters/AwsRegion'
      responses:
        '200':
          description: EBS volume types and local pricing availability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EbsTypeCatalog'
        default:
          $ref: '#/components/responses/Problem'
...
  /pricing/aws/resolve:
    post:
      operationId: resolveAwsPrices
      tags: [Pricing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceResolutionRequest'
      responses:
        '200':
          description: AWS snapshot status, including normal unavailable outcomes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResolution'
        default:
          $ref: '#/components/responses/Problem'
  /pricing/aws/refresh:
    post:
      operationId: refreshAwsPrices
      tags: [Pricing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceResolutionRequest'
      responses:
        '200':
          description: Forced AWS live-resolution outcome.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResolution'
        default:
          $ref: '#/components/responses/Problem'
  /pricing/azure/resolve:
    post:
      operationId: resolveAzurePrices
      tags: [Pricing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceResolutionRequest'
      responses:
        '200':
          description: Azure snapshot status, including normal unavailable outcomes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResolution'
        default:
          $ref: '#/components/responses/Problem'
  /pricing/azure/refresh:
    post:
      operationId: refreshAzurePrices
      tags: [Pricing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceResolutionRequest'
      responses:
        '200':
          description: Forced Azure live-resolution outcome.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResolution'
        default:
          $ref: '#/components/responses/Problem'
  /calculations:
    post:
      operationId: calculateProject
      tags: [Calculations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculationRequest'
      responses:
        '200':
          description: Server-owned calculation revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationRevision'
        default:
          $ref: '#/components/responses/Problem'
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '200':
          description: Projects owned by the current principal.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableProject'
      responses:
        '201':
          description: Created owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
  /projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '200':
          description: Owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      parameters:
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableProject'
      responses:
        '200':
          description: Updated owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '204':
          description: Project deleted immediately.
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    containerAppsAuth:
      type: openIdConnect
      openIdConnectUrl: https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration
      description: Authentication is enforced by Azure Container Apps built-in authentication.
  parameters:
    AwsRegion:
      name: region
      in: query
      required: true
      schema:
        type: string
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 Problem Details response.
      headers:
        Retry-After:
          description: Seconds to wait when the request quota is exceeded.
          schema:
            type: string
        ETag:
          description: Current project ETag after a failed precondition.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Decimal:
      type: string
      pattern: '^-?(0|[1-9][0-9]*)(\.[0-9]+)?$'
      examples: ['8760', '0.10', '12345.67']
    ResolutionStatus:
      type: string
      enum: [fresh, cached, stale, unavailable]
...
    Ec2Instance:
      type: object
      required: [instance_type, source_vcpu, memory_gib]
      properties:
        instance_type:
          type: string
        source_vcpu:
          type: integer
          minimum: 1
        memory_gib:
          $ref: '#/components/schemas/Decimal'
    RdsInstance:
      $ref: '#/components/schemas/Ec2Instance'
    RdsOption:
      type: object
      required: [deployment, commercial_term, storage_class]
      properties:
        deployment:
          type: string
          enum: [single_az, multi_az]
        commercial_term:
          type: string
        storage_class:
          type: string
    EbsType:
      type: object
      required: [key, label, price_required, pricing_available]
      properties:
        key:
          type: string
          enum: [gp3, io2, ephemeral]
        label:
          type: string
        price_required:
          type: boolean
        pricing_available:
          type: boolean
    Ec2InstanceCatalog:
      allOf:
        - $ref: '#/components/schemas/CatalogMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Ec2Instance'
    RdsInstanceCatalog:
      allOf:
        - $ref: '#/components/schemas/CatalogMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RdsInstance'
    RdsOptionCatalog:
      allOf:
        - $ref: '#/components/schemas/CatalogMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RdsOption'
    EbsTypeCatalog:
      allOf:
        - $ref: '#/components/schemas/CatalogMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/EbsType'
    CatalogMetadata:
      type: object
      required: [status, currency, retrieved_at, source_urls, warnings]
      properties:
        status:
          $ref: '#/components/schemas/ResolutionStatus'
        currency:
          type: string
          const: USD
        retrieved_at:
          type: [string, 'null']
          format: date-time
        source_urls:
          type: array
          items:
            type: string
            format: uri
        warnings:
          type: array
          items:
            type: string
...
    PurchaseOptionKey:
      type: string
      enum: [payg, ahb, one-year, ahbone-year, three-year, ahbthree-year, sv-one-year, ahbsv-one-year]
    PriceResolutionRequest:
      type: object
      additionalProperties: false
      required: [currency, azure_region, resources]
      properties:
        currency:
          type: string
          const: USD
        aws_region:
          type: [string, 'null']
        azure_region:
          type: string
        resources:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Resource'
    PriceResolution:
      type: object
      required: [provider, status, snapshot_id, warnings]
      properties:
        provider:
          type: string
          enum: [aws, azure]
        status:
          $ref: '#/components/schemas/ResolutionStatus'
        snapshot_id:
          type: [string, 'null']
        retrieved_at:
          type: [string, 'null']
          format: date-time
        warnings:
          type: array
          items:
            type: string
    CalculationRequest:
      allOf:
        - $ref: '#/components/schemas/EditableProjectFields'
        - type: object
          properties:
            expected_formula_version:
              type: [string, 'null']
      unevaluatedProperties: false
    CalculationRevision:
      type: object
      required: [formula_version, aws_snapshot_id, azure_snapshot_id, resource_results, portfolio_totals, warnings]
      properties:
        formula_version:
          type: string
        aws_snapshot_id:
          type: [string, 'null']
        azure_snapshot_id:
          type: [string, 'null']
        resource_results:
          type: array
          items:
            $ref: '#/components/schemas/ResourceCalculation'
        portfolio_totals:
          $ref: '#/components/schemas/PortfolioTotals'
        warnings:
          type: array
          items:
            type: string
    ResourceCalculation:
      type: object
      required:
        - resource_id
        - mapping_status
        - aws_pricing_status
        - azure_pricing_status
        - target_selection
        - source_costs
        - azure_costs
        - savings
        - explanation_steps
        - unresolved_components
      properties:
        resource_id:
          type: string
          format: uuid
        mapping_status:
          type: [string, 'null']
          enum: [mapped, no_mapping, null]
        aws_pricing_status:
          $ref: '#/components/schemas/PricingStatus'
        azure_pricing_status:
          $ref: '#/components/schemas/PricingStatus'
        target_selection:
          oneOf:
            - $ref: '#/components/schemas/TargetSelection'
            - type: 'null'
        source_costs:
          oneOf:
            - $ref: '#/components/schemas/SourceCostBreakdown'
            - type: 'null'
        azure_costs:
          oneOf:
            - $ref: '#/components/schemas/AzureCostBreakdown'
            - type: 'null'
        savings:
          oneOf:
            - $ref: '#/components/schemas/SavingsBreakdown'
            - type: 'null'
        explanation_steps:
          type: array
          items:
            $ref: '#/components/schemas/ExplanationStep'
        unresolved_components:
          type: array
          items:
            $ref: '#/components/schemas/UnresolvedComponent'
    PricingStatus:
      type: string
      enum: [fresh, cached, stale, unavailable, not_required]
    PortfolioTotals:
      type: object
      required:
        - aws_all_rows_total
        - aws_mapped_rows_total
        - azure_mapped_rows_total
        - required_portfolio_adjustment
        - selected_parity_adjustment
        - portfolio_after_selected_parity
        - portfolio_difference
        - comparable_resource_count
        - no_mapping_resource_count
        - price_unavailable_resource_count
      properties:
        aws_all_rows_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        aws_mapped_rows_total:
          $ref: '#/components/schemas/Decimal'
        azure_mapped_rows_total:
          $ref: '#/components/schemas/Decimal'
        required_portfolio_adjustment:
          $ref: '#/components/schemas/Decimal'
        selected_parity_adjustment:
          $ref: '#/components/schemas/Decimal'
        portfolio_after_selected_parity:
          $ref: '#/components/schemas/Decimal'
        portfolio_difference:
          $ref: '#/components/schemas/Decimal'
        comparable_resource_count:
          type: integer
          minimum: 0
        no_mapping_resource_count:
          type: integer
          minimum: 0
        price_unavailable_resource_count:
          type: integer
          minimum: 0
    SourceCostBreakdown:
      type: object
      required: [compute_gross, compute_net, license_gross, license_net, storage_gross, storage_net, hardware_annual, electricity_annual, total]
      properties:
        compute_gross:
          $ref: '#/components/schemas/Decimal'
        compute_net:
          $ref: '#/components/schemas/Decimal'
        license_gross:
          $ref: '#/components/schemas/Decimal'
        license_net:
          $ref: '#/components/schemas/Decimal'
        storage_gross:
          $ref: '#/components/schemas/Decimal'
        storage_net:
          $ref: '#/components/schemas/Decimal'
        hardware_annual:
          $ref: '#/components/schemas/Decimal'
        electricity_annual:
          $ref: '#/components/schemas/Decimal'
        total:
          $ref: '#/components/schemas/Decimal'
    AzureCostBreakdown:
      type: object
      required: [compute_gross, additional_ram_gb, additional_ram_gross, compute_plus_ram_net, license_gross, license_net, storage_gross, storage_net, total_before_parity]
      properties:
        compute_gross:
          $ref: '#/components/schemas/Decimal'
        additional_ram_gb:
          $ref: '#/components/schemas/Decimal'
        additional_ram_gross:
          $ref: '#/components/schemas/Decimal'
        compute_plus_ram_net:
          $ref: '#/components/schemas/Decimal'
        license_gross:
          $ref: '#/components/schemas/Decimal'
        license_net:
          $ref: '#/components/schemas/Decimal'
        storage_gross:
          $ref: '#/components/schemas/Decimal'
        storage_net:
          $ref: '#/components/schemas/Decimal'
        total_before_parity:
          $ref: '#/components/schemas/Decimal'
...
    TargetSelection:
      type: object
      required: [mapping_status, requested_tier, nggp_iops_limit, selected, candidates, outcome_reasons, storage_escalation]
      properties:
        mapping_status:
          type: string
          enum: [mapped, no_mapping]
        requested_tier:
          $ref: '#/components/schemas/ServiceTier'
        nggp_iops_limit:
          type: integer
          minimum: 0
        selected:
          oneOf:
            - $ref: '#/components/schemas/SelectedTarget'
            - type: 'null'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/CandidateEvaluation'
        outcome_reasons:
          type: array
          items:
            $ref: '#/components/schemas/SelectionReason'
        storage_escalation:
          oneOf:
            - $ref: '#/components/schemas/StorageEscalation'
            - type: 'null'
...
    ExplanationStep:
      type: object
      required: [code, message, values]
      properties:
        code:
          type: string
        message:
          type: string
        values:
          type: object
          additionalProperties:
            type: string
...
    EditableProject:
      allOf:
        - $ref: '#/components/schemas/EditableProjectFields'
      unevaluatedProperties: false
    EditableProjectFields:
      type: object
      required: [name, settings, resources, aws_price_snapshot_id, azure_price_snapshot_id]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, 'null']
          maxLength: 500
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        resources:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Resource'
        aws_price_snapshot_id:
          type: [string, 'null']
        azure_price_snapshot_id:
          type: [string, 'null']
    ProjectSettings:
      type: object
      additionalProperties: false
      required:
        - project_type
        - aws_region
        - azure_region
        - currency
        - source_compute_discount
        - source_license_discount
        - source_storage_discount
        - azure_compute_discount
        - azure_license_discount
        - azure_storage_discount
        - selected_parity_adjustment
        - default_annual_hours
        - default_mi_purchase_option
      properties:
        project_type:
          type: string
          enum: [ec2, rds, on_prem]
        aws_region:
          type: [string, 'null']
        azure_region:
          type: string
        currency:
          type: string
          const: USD
        source_compute_discount:
          $ref: '#/components/schemas/Decimal'
        source_license_discount:
          $ref: '#/components/schemas/Decimal'
        source_storage_discount:
          $ref: '#/components/schemas/Decimal'
        azure_compute_discount:
          $ref: '#/components/schemas/Decimal'
        azure_license_discount:
          $ref: '#/components/schemas/Decimal'
        azure_storage_discount:
          $ref: '#/components/schemas/Decimal'
        selected_parity_adjustment:
          $ref: '#/components/schemas/Decimal'
        default_annual_hours:
          $ref: '#/components/schemas/Decimal'
        default_mi_purchase_option:
          $ref: '#/components/schemas/PurchaseOptionKey'
        enterprise_license_sa_usd_per_two_core_pack:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        standard_license_sa_usd_per_two_core_pack:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        remaining_coverage_months:
          type: [integer, 'null']
          enum: [12, 24, 36, null]
        electricity_rate_usd_per_kwh:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    Resource:
      oneOf:
        - $ref: '#/components/schemas/Ec2Resource'
        - $ref: '#/components/schemas/RdsResource'
        - $ref: '#/components/schemas/OnPremResource'
      discriminator:
        propertyName: source_type
...
    Ec2Resource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, instance_type, volumes]
          properties:
            source_type:
              type: string
              const: ec2
            instance_type:
              type: string
            volumes:
              type: array
              minItems: 1
              maxItems: 50
              items:
                $ref: '#/components/schemas/EbsVolume'
...
    RdsResource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, instance_type, deployment, commercial_term, storage_class, source_max_iops]
          properties:
            source_type:
              type: string
              const: rds
            instance_type:
              type: string
            deployment:
              type: string
              enum: [single_az, multi_az]
            commercial_term:
              type: string
            storage_class:
              type: string
            source_max_iops:
              type: integer
              minimum: 0
              maximum: 1000000000
    OnPremResource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, source_vcpu, licensable_cores, source_max_iops, hardware_capex_usd, depreciation_years]
          properties:
            source_type:
              type: string
              const: on_prem
            source_vcpu:
              type: integer
              minimum: 1
              maximum: 100000
            licensable_cores:
              type: integer
              minimum: 1
              maximum: 100000
            source_max_iops:
              type: integer
              minimum: 0
              maximum: 1000000000
            hardware_capex_usd:
              $ref: '#/components/schemas/Decimal'
            depreciation_years:
              $ref: '#/components/schemas/Decimal'
            average_power_kw_override:
              oneOf:
                - $ref: '#/components/schemas/Decimal'
                - type: 'null'
    Project:
      allOf:
        - $ref: '#/components/schemas/EditableProjectFields'
        - type: object
          required: [id, created_at, updated_at, formula_version, schema_version, latest_calculation_revision]
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            formula_version:
              type: string
            schema_version:
              type: string
            latest_calculation_revision:
              oneOf:
                - $ref: '#/components/schemas/CalculationRevision'
                - type: 'null'
      unevaluatedProperties: false
    ProjectSummary:
      type: object
      required: [id, name, project_type, modified_at, source_region, azure_region, resource_count, source_annual_total, azure_annual_total]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project_type:
          type: string
          enum: [ec2, rds, on_prem]
        modified_at:
          type: string
          format: date-time
        source_region:
          type: [string, 'null']
        azure_region:
          type: string
        resource_count:
          type: integer
        source_annual_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        azure_annual_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    Problem:
      type: object
      required: [type, title, status, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [pointer, code, message]
            properties:
              pointer:
                type: string
              code:
                type: string
              message:
                type: string